# docs/swagger.yaml
basePath: /
definitions:
  common.ErrorResponse:
    properties:
      details:
        additionalProperties: true
        type: object
      error_code:
        type: string
      message:
        type: string
    type: object
...
  httphandlers.AsyncData:
    properties:
      analysisID:
        type: string
      ipAddr:
        type: string
      subdomain:
        type: string
    type: object
  httphandlers.ExternalIDResp:
    properties:
      external_id:
        example: bb52aefb-e021-4ece-89e5-fd73ce30643c
        type: string
    type: object
...
  httphandlers.TerminateAllResponse:
    properties:
      batch:
        items:
          type: string
        type: array
      failed_batch:
        items:
          type: string
        type: array
      failed_vice:
        items:
          type: string
        type: array
      vice:
        items:
          type: string
        type: array
    type: object
...
host: localhost:60000
info:
  contact: {}
  description: The app-exposer API for the Discovery Environment's VICE feature.
  license:
    name: 3-Clause BSD License
    url: https://github.com/cyverse-de/app-exposer?tab=License-1-ov-file#readme
  title: app-exposer
  version: "1.0"
paths:
...
  /vice/{id}/exit:
    post:
      description: |-
        Terminates the VICE analysis deployment and cleans up
        resources asscociated with it. Does not save outputs first. Uses
        the external-id label to find all of the objects in the configured
        namespace associated with the job. Deletes the following objects:
        HTTP routes, services, deployments, and configmaps.
      operationId: exit
      parameters:
      - description: The external ID of the VICE analysis
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Terminates a VICE analysis
  /vice/{id}/save-and-exit:
    post:
      description: |-
        Handles requests to save the output files in iRODS and then exit.
        The exit portion will only occur if the save operation succeeds. The operation is
        performed inside of a goroutine so that the caller isn't waiting for hours/days for
        output file transfers to complete.
      operationId: save-and-exit
      parameters:
      - description: external ID of the analysis
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Save files and terminate VICE analysis
...
  /vice/admin/analyses/{}/save-and-exit:
    post:
      description: |-
        Handles requests to save the output files in iRODS and
        then exit. This version of the call operates based on the analysis ID and does
        not require user information to be required by the caller. Otherwise, the docs
        for the VICESaveAndExit function apply here as well.
      operationId: admin-save-and-exit
      parameters:
      - description: Analysis ID
        in: path
        name: analysis-id
        required: true
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Admin endpoint to trigger output file transfer and analysis exit
...
  /vice/admin/analyses/{analysis-id}/exit:
    post:
      description: |-
        Terminates the VICE analysis based on the analysisID and
        and should not require any user information to be provided. Otherwise, the
        documentation for VICEExit applies here as well.
      operationId: admin-exit
      parameters:
      - description: The external ID of the VICE analysis
        in: path
        name: analysis-id
        required: true
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Terminates a VICE analysis
  /vice/admin/analyses/{analysis-id}/external-id:
    get:
      description: |-
        Returns the external ID associated with the provided analysis ID.
        Only returns the first external ID in multi-step analyses.
      operationId: admin-get-external-id
      parameters:
      - description: analysis UUID
        in: path
        maxLength: 36
        minLength: 36
        name: analysis-id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httphandlers.ExternalIDResp'
        "400":
          description: id parameter is empty
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Returns external ID
...
  /vice/admin/terminate-all:
    post:
      description: |-
        Terminates all analyses marked as running in the DE database.
        Does not check for analyses in the cluster but not marked as running in the database.
        Terminates both VICE and batch analyses.
      operationId: terminate-all-analyses
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httphandlers.TerminateAllResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Terminates all analyses
  /vice/apply-labels:
    post:
      description: |-
        Asynchronously applies labels to all running VICE analyses.
        The application of the labels may not be complete by the time the response is returned.
      operationId: apply-async-labels
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Applies labels to running VICE analyses.
  /vice/async-data:
    get:
      description: |-
        Returns data that is applied to analyses outside of an API call.
        The returned data is not returned asynchronously, despite the name of the call.
      operationId: async-data
      parameters:
      - description: External ID
        in: query
        name: external-id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httphandlers.AsyncData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Returns data that is generated asynchronously from the job launch.
...
swagger: "2.0"